Namewspaced service classes
