+++ spec/services/api/v1/data/emission_pathways/csv_content_spec.rb
@@ -1,13 +1,13 @@
 require 'rails_helper'
 
-RSpec.describe Api::V1::Data::EmissionPathwaysCsvContent do
+RSpec.describe Api::V1::Data::EmissionPathways::CsvContent do
   include_context 'emission pathways'
 
   describe :call do
     it 'Replaces nulls with N/A' do
-      filter = Api::V1::Data::EmissionPathwaysFilter.new({})
+      filter = Api::V1::Data::EmissionPathways::Filter.new({})
       parsed_csv = CSV.parse(
-        Api::V1::Data::EmissionPathwaysCsvContent.new(filter).call
+        Api::V1::Data::EmissionPathways::CsvContent.new(filter).call
       )
       expect(parsed_csv.last.last).to eq('N/A')
     end
